armel (linux) compiler complaint fixed
part locked state is saved and restored

git-svn-id: http://fritzing.googlecode.com/svn/trunk@5436 85f94c43-f63a-0410-9d22-21439860c375

+++ fritzing/src/partsbinpalette/graphicsflowlayout.cpp
@@ -80,7 +80,8 @@ int GraphicsFlowLayout::doLayout(const QRectF &rect) {
 		item->setGeometry(QRectF(QPoint(x, y), item->preferredSize()));
 
 		x = nextX;
-		lineHeight = qMax(lineHeight, item->preferredSize().height());
+		// item->preferredSize().height() returns qreal, armel compiler complains
+		lineHeight = qMax(lineHeight, (double) item->preferredSize().height());
 	}
 
 	m_lastWidth = rect.width();
